refactor:전시회 상세 총 좋아요 수 낙관적 업데이트로 변경
